Move thread local thread/pool numbers back to cpp file
